Initial implementation of dispatcihng (untested)
