+++ distance_metrics/src/abs_double.hpp
@@ -1,0 +1,6 @@
+#include "abs.hpp"
+#include "xsimd/xsimd.hpp"
+
+namespace xs = xsimd;
+
+template double_batch xsimd_abs::operator()<double_batch, xs::sse3, double>(xs::sse3, const double_batch&);
