+++ distance_metrics/src/abs_float.hpp
@@ -1,0 +1,6 @@
+#include "abs.hpp"
+#include "xsimd/xsimd.hpp"
+
+namespace xs = xsimd;
+
+template float_batch xsimd_abs::operator()<float_batch, xs::sse3, float>(xs::sse3, const float_batch&);
